+++ oldtestfiles/graphConstructorBasic.py
