+++ oldtestfiles/graphConstructorGPT.py
