--- test.py
@@ -1,3 +1,0 @@
-from mesh import load_result, get_parser 
-
-load_result("/Users/vikyath/SBU/Robotics Lab/conceptGraph/graphGenerator/detections/room1.pkl.gz")
